+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Improve accessibility of `<ad-block>`s (contrast, keyboard navigation)
+- Fix parsing of params in `<form is="share-target">`
 
 ## [v1.0.11] - 2020-10-02